+++ libraries/kotlinx-metadata/jvm/ChangeLog.md
@@ -7,6 +7,8 @@
     - Breaking change: JvmPropertyExtensionVisitor.visit has a new parameter `jvmFlags: Flags`
 - Correctly write "null" constant value in effect expression of a contract
 - Rename `desc` parameters to `signature` in JvmFunctionExtensionVisitor, JvmPropertyExtensionVisitor, JvmConstructorExtensionVisitor
+- Do not expose KmExtensionType internals
+- Add KmExtensionVisitor.type to get dynamic type of an extension visitor
 
 ## 0.0.5